+++ packages/appcd-subprocess/CHANGELOG.md
@@ -4,6 +4,7 @@
  * fix: Properly handle IPC send errors when child process goes away.
  * fix: Kill non-detached child processes when response stream is closed by the caller.
  * fix: Decouple subprocess data from `/appcd/status`.
+ * fix: Hide child process event emitter so that it won't leak when subprocess info is stringified.
 
 # v4.0.0 (Jun 12, 2020)